move utils to openpmd_helper file
